+++ protobuf/src/main/java/com/fasterxml/jackson/dataformat/protobuf/schemagen/AnnotationBasedTagGenerator.java
@@ -8,7 +8,7 @@ public class AnnotationBasedTagGenerator implements TagGenerator
     public int nextTag(BeanProperty writer) {
         Integer ix = writer.getMetadata().getIndex();
         if (ix != null) {
-            return ix.intValue();
+            return ix;
         }
         throw new IllegalStateException("No index metadata found for " + writer.getFullName()
                 + " (usually annotated with @JsonProperty.index): either annotate all properties of type " + writer.getWrapperName().getSimpleName() + " with indexes or none at all");
